Add skip if key instructions
Adds the skip if key instruction. Macros were added in defs.h to ease
the the implementation of the callback for the frontend. Furthermore the
callback was changed to return a 32 bit unsigned integer instead of a 16
bit since all keys wouldn't fit in a 16 bit bitmask.

+++ include/private/crisp8_private.h
@@ -18,7 +18,7 @@
 #include "config.h"
 
 typedef void (*crisp8AudioCallback) (void);
-typedef uint16_t (*crisp8InputCallback) (void);
+typedef uint32_t (*crisp8InputCallback) (void);
 
 struct chip8_s
 {
